fundy: device initial setup.
this is the device profile for 7278B1 (also supports 7278B0), aka
"fundy" device.  the fundy device is not yet fully functional for
this reason we do not yet provide binaries for it.  at the moment
this is therefore for reference; but can be built using full source
distribution (ie if kernel, bootloader and gcc are available).

Change-Id: I956e37e56cb87cdf253616eb95ceba98b725734c
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>