fundy: move to v7-64 memory address layout.
for 7278, the v7-64 memory layout is recommended and in fact
enforced for productions parts.  as such the current v7-32 memory
layout implemented by the 'fundy' device is not appropriate,
even more so since the device is used in a mixed environment of
production and development parts with a single image.

therefore, we now switch to v7-64 layout support only and adjust
the memory and heaps configuration accordingly.

however, specifically on the 7278IPA device using development parts,
we need to apply a small board change to ensure the v7-64 layout is
selected by default, this is the less of all evils solution for this.

the change is to cut R901 and connect the line to GND.  this change
causes the activity LED on the built in ETH to no longer work either,
a small side effect of the change.

refs #SWANDROID-5294

Change-Id: I2fafad850e471f23753ce43c7ff753f7f29e4801
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>